refactor: move index html, css, js to separate files
